+++ CustomTextualInputOutputPort.h
@@ -65,7 +65,7 @@ public:
     bool hasPosition() const;
     bool hasSetPosition() const;
     Object position() const;
-    bool setPosition(int position);
+    bool setPosition(int64_t position);
     int close();
     bool isClosed() const;
 private:
