+++ CustomTextualInputPort.h
@@ -49,7 +49,7 @@ public:
     bool hasPosition() const;
     bool hasSetPosition() const;
     Object position() const;
-    bool setPosition(int position);
+    bool setPosition(int64_t position);
 
     // Port interface
     ucs4string toString();
