+++ StringTextualInputPort.h
@@ -50,7 +50,7 @@ public:
     bool hasPosition() const;
     bool hasSetPosition() const;
     Object position() const;
-    bool setPosition(int position);
+    bool setPosition(int64_t position);
     int getLineNo() const;
     Transcoder* transcoder() const;
